+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.0] - 2024-12-06
+## [0.1.0] - 2024-12-06
 
 ### Added
 - Library target structure (`AWContextLib`) for better modularity
@@ -30,7 +30,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Concurrent file access issues in tests
 - Missing public modifiers on library types
 
-## [1.0.0] - 2024-12-05
+### Changed
+- Query command now accepts date as first positional argument
+- Query command --start and --end are now optional overrides
+
+## [0.0.1] - 2024-12-05
 
 ### Added
 - Initial release